Replace License header by SPDX-License-Identifier (BSD-2-Clause)

+++ src/breath.h
@@ -1,30 +1,9 @@
 /*
+ * SPDX-License-Identifier: BSD-2-Clause
+ *
  * Copyright (c)2014-2022, Luc Hondareyte
- * 
  * All rights reserved.
  * 
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- * 
- *  * Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- *  * Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- * 
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER
- * OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
- * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
- * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
- * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
- * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
- * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
  */
 
 #ifndef __BREATH_H__
@@ -34,7 +13,7 @@
 #define MAX_BREATH_MODE         6
 
 /* Linear curve */
-static const int a_curve[TABLE_SIZE] = {
+static const uint8_t a_curve[TABLE_SIZE] = {
 
 	0, 1, 2, 3, 4, 5, 6, 7, 
 	8, 9, 10, 11, 12, 13, 14, 15, 
@@ -55,7 +34,7 @@ static const int a_curve[TABLE_SIZE] = {
 };
 
 /* pseudo exponential curve */
-static const int b_curve[TABLE_SIZE] = {
+static const uint8_t b_curve[TABLE_SIZE] = {
 
 	1, 1, 2, 3, 4, 5, 6, 7, 
 	8, 9, 10, 11, 12, 13, 14, 15, 
@@ -78,7 +57,7 @@ static const int b_curve[TABLE_SIZE] = {
 /*  Curve that simulate WT11 response */
 
 #if defined __WT11_COMPAT__
-static const int c_curve[TABLE_SIZE] = {
+static const uint8_t c_curve[TABLE_SIZE] = {
 
 	20, 20, 20, 20, 20, 20, 20, 20, 
 	20, 20, 20, 20, 21, 21, 21, 21, 
@@ -98,7 +77,7 @@ static const int c_curve[TABLE_SIZE] = {
 	117, 118, 120, 122, 123, 125, 126, 127
 };
 
-static const int d_curve[TABLE_SIZE] = {
+static const uint8_t d_curve[TABLE_SIZE] = {
 
 	25, 25, 25, 25, 25, 25, 25, 25,
 	25, 25, 25, 25, 25, 26, 26, 26,
@@ -118,7 +97,7 @@ static const int d_curve[TABLE_SIZE] = {
 	116, 117, 119, 120, 122, 123, 125, 126 
 };
 
-static const int e_curve[TABLE_SIZE] = {
+static const uint8_t e_curve[TABLE_SIZE] = {
 
 	30, 30, 30, 30, 30, 30, 30, 30, 
 	30, 30, 30, 30, 30, 31, 31, 31, 
@@ -139,7 +118,7 @@ static const int e_curve[TABLE_SIZE] = {
 };	
 
 #else
-static const int c_curve[TABLE_SIZE] = {
+static const uint8_t c_curve[TABLE_SIZE] = {
 
 	0, 3, 6, 9, 12, 15, 18, 21, 
 	24, 27, 30, 33, 36, 39, 42, 45, 
@@ -159,7 +138,7 @@ static const int c_curve[TABLE_SIZE] = {
 	124, 124, 124, 125, 125, 125, 126, 126
 };
 
-static const int d_curve[TABLE_SIZE] = {
+static const uint8_t d_curve[TABLE_SIZE] = {
 
 	0, 3, 6, 9, 12, 15, 18, 21, 
 	24, 27, 30, 33, 36, 39, 42, 45, 
@@ -179,7 +158,7 @@ static const int d_curve[TABLE_SIZE] = {
 	124, 125, 125, 125, 126, 126, 127, 127
 };
 
-static const int e_curve[TABLE_SIZE] = {
+static const uint8_t e_curve[TABLE_SIZE] = {
 
 	0, 1, 1, 1, 1, 2, 2, 3, 
 	3, 3, 4, 4, 5, 5, 6, 6, 
@@ -200,7 +179,7 @@ static const int e_curve[TABLE_SIZE] = {
 };
 #endif
 
-static const int f_curve[TABLE_SIZE] = {
+static const uint8_t f_curve[TABLE_SIZE] = {
 
 	0, 0, 0, 1, 1, 2, 2, 3, 
 	3, 3, 4, 4, 5, 5, 6, 6,
